requirements.txt
requests
python-dotenv
rich
questionary
pypdf
